Change functions an name from index.js  to contacts.js
